+++ packages/api/src/promise/types.ts
@@ -2,10 +2,10 @@
 // This software may be modified and distributed under the terms
 // of the Apache-2.0 license. See the LICENSE file for details.
 
-import { QueryableStorageFunction as QueryableStorageFunctionBase, SubmittableExtrinsicFunction as SubmittableExtrinsicFunctionBase } from '../types';
+import { QueryableStorageEntry as QueryableStorageEntryBase, SubmittableExtrinsicFunction as SubmittableExtrinsicFunctionBase } from '../types';
 
 import { SubmittableExtrinsic as SubmittableExtrinsicBase } from '../SubmittableExtrinsic';
 
-export type QueryableStorageFunction = QueryableStorageFunctionBase<'promise'>;
+export type QueryableStorageEntry = QueryableStorageEntryBase<'promise'>;
 export type SubmittableExtrinsic = SubmittableExtrinsicBase<'promise'>;
 export type SubmittableExtrinsicFunction = SubmittableExtrinsicFunctionBase<'promise'>;